requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
faiss-cpu==1.8.0.post1
redis==5.1.1
aiofiles==24.1.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pypdf==4.3.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
line-profiler==4.1.3
